requirements.txt
appnope==0.1.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
cycler==0.10.0
dask==0.8.1
decorator==4.0.9
docopt==0.6.2
ecdsa==0.13
Fabric==1.10.2
functools32==3.2.3.post2
gnureadline==6.3.3
h5py==2.6.0
ipykernel==4.3.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
jupyter-themes==0.1
Keras==0.3.3
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.7.2
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.11
nose==1.3.7
notebook==4.1.0
numpy==1.10.4
paramiko==1.16.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.1
protobuf==3.0.0b2
ptyprocess==0.5.1
pycrypto==2.6.1
Pygments==2.1.3
pyparsing==2.1.0
pypng==0.0.18
python-dateutil==2.5.0
pytz==2015.7
PyYAML==3.11
pyzmq==15.2.0
qtconsole==4.2.0
scikit-image==0.12.3
scipy==0.17.0
selenium==2.52.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
tensorflow==0.7.1
terminado==0.6
Theano==0.8.1
toolz==0.7.4
tornado==4.3
tqdm==3.8.0
traitlets==4.1.0
